requirements.txt
# Production AI Companion Requirements - PRODUCTION ONLY

# Core Flask and Web Framework
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7

# Environment and Configuration
python-dotenv==1.0.0
python-decouple==3.8

# HTTP and Web Scraping
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1
beautifulsoup4==4.12.2
lxml==4.9.3
readability==0.3.1

# Selenium for JavaScript rendering
selenium==4.15.2
webdriver-manager==4.0.1

# Google Cloud and Firestore
google-cloud-firestore==2.13.1
google-oauth2-tool==0.0.3
google-auth==2.24.0
google-cloud-core==2.3.3

# AI Provider SDKs
openai==1.3.7
anthropic==0.8.1
# google-generativeai==0.3.2  # Uncomment if using Gemini
# azure-cognitiveservices-language-textanalytics==5.2.0  # For Microsoft

# OAuth and Authentication
authlib==1.2.1
oauthlib==3.2.2
PyJWT==2.8.0

# Security and Encryption
cryptography==41.0.7
bcrypt==4.1.2

# Data Processing and Analysis
pandas==2.1.4
numpy==1.25.2
python-dateutil==2.8.2

# Async and Concurrency
asyncio==3.4.3
aiofiles==23.2.1

# Logging and Monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Database and Storage
SQLAlchemy==2.0.23
redis==5.0.1

# Production Server
gunicorn==21.2.0
gevent==23.9.1

# System and Performance
psutil==5.9.6
memory-profiler==0.61.0

# Security
flask-limiter==3.5.0
flask-talisman==1.1.0